ARUHA-56: extracted json mapper bean creation to separate config class;
