+++ src/test/java/de/zalando/aruha/nakadi/controller/EventPublishingControllerTest.java
@@ -1,8 +1,8 @@
 package de.zalando.aruha.nakadi.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import de.zalando.aruha.nakadi.config.JsonConfig;
 import de.zalando.aruha.nakadi.exceptions.NakadiException;
-import de.zalando.aruha.nakadi.config.NakadiConfig;
 import de.zalando.aruha.nakadi.domain.EventType;
 import de.zalando.aruha.nakadi.repository.DuplicatedEventTypeNameException;
 import de.zalando.aruha.nakadi.repository.EventTypeRepository;
@@ -43,8 +43,8 @@ public class EventPublishingControllerTest {
 
     private final MockMvc mockMvc;
 
-    public EventPublishingControllerTest() throws NakadiException, DuplicatedEventTypeNameException {
-        final ObjectMapper objectMapper = new NakadiConfig().jacksonObjectMapper();
+    public EventPublishingControllerTest() throws NakadiException {
+        final ObjectMapper objectMapper = new JsonConfig().jacksonObjectMapper();
 
         jsonHelper = new JsonTestHelper(objectMapper);
         topicRepository.createTopic(EVENT_TYPE_WITH_TOPIC);
