+++ src/test/java/de/zalando/aruha/nakadi/controller/PartitionsControllerTest.java
@@ -2,7 +2,7 @@ package de.zalando.aruha.nakadi.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableList;
-import de.zalando.aruha.nakadi.config.NakadiConfig;
+import de.zalando.aruha.nakadi.config.JsonConfig;
 import de.zalando.aruha.nakadi.domain.TopicPartition;
 import de.zalando.aruha.nakadi.exceptions.NakadiException;
 import de.zalando.aruha.nakadi.exceptions.ServiceUnavailableException;
@@ -53,7 +53,7 @@ public class PartitionsControllerTest {
 
     @Before
     public void before() {
-        final ObjectMapper objectMapper = new NakadiConfig().jacksonObjectMapper();
+        final ObjectMapper objectMapper = new JsonConfig().jacksonObjectMapper();
         jsonHelper = new JsonTestHelper(objectMapper);
 
         topicRepositoryMock = mock(TopicRepository.class);
